header and first page completed
in html boiler code has been made using a vague idea of the website design sketch

**for the header**
the whole header is considered as a flex container so that i can set the margin to -1 to fill the space in with color.

position sticky is the most important here as i made it to stick on top of things to overlap everything else (but not while scrolling)
for scrolling i used z index with large value 999 to layer it on top of elements as x,y,"z" is used for 3d position.

also i made both the left right items on two divs as each is treated as a flex

**page one**
for the first page the container is flex and position is relative as i am setting things below that, that is absolute and it should be relative to the container dimenasions.

for the text that overlaps the image i gave the position as absolute and the values as top 50% and left 50% which is the center of the container whose center will change relatively

i have used transform translate tag as the text was not centered properly bacause of its parent containers relative position